batcheval,storage: move EndTransaction to batcheval
Move the last remaining command, EndTransaction, to the batcheval
package, where it logically belongs. In doing so, resolve a longstanding
TODO.

The one-phase commit logic still lives in storage and, unfortunately,
ends up tightly coupled with the EndTransaction machinery in batcheval,
but this is a step in the right direction. By contrast, the extraction
of writeInitial[Replica]State into the stateloader package goes quite
smoothly.

Release note: None
