+++ pkg/storage/stateloader/initial.go
@@ -12,36 +12,38 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package storage
+package stateloader
 
 import (
 	"context"
 
-	opentracing "github.com/opentracing/opentracing-go"
-	"github.com/pkg/errors"
-
-	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
-	"github.com/cockroachdb/cockroach/pkg/storage/abortspan"
-	"github.com/cockroachdb/cockroach/pkg/storage/batcheval"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
-	"github.com/cockroachdb/cockroach/pkg/storage/stateloader"
 	"github.com/cockroachdb/cockroach/pkg/storage/storagebase"
-	"github.com/cockroachdb/cockroach/pkg/storage/txnwait"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/pkg/errors"
 )
 
-// writeInitialReplicaState sets up a new Range, but without writing an
+// raftInitialLog{Index,Term} are the starting points for the raft log. We
+// bootstrap the raft membership by synthesizing a snapshot as if there were
+// some discarded prefix to the log, so we must begin the log at an arbitrary
+// index greater than 1.
+const (
+	raftInitialLogIndex = 10
+	raftInitialLogTerm  = 5
+)
+
+// WriteInitialReplicaState sets up a new Range, but without writing an
 // associated Raft state (which must be written separately via
 // synthesizeRaftState before instantiating a Replica). The main task is to
 // persist a ReplicaState which does not start from zero but presupposes a few
 // entries already having applied. The supplied MVCCStats are used for the Stats
 // field after adjusting for persisting the state itself, and the updated stats
 // are returned.
-func writeInitialReplicaState(
+func WriteInitialReplicaState(
 	ctx context.Context,
 	st *cluster.Settings,
 	eng engine.ReadWriter,
@@ -51,7 +53,7 @@ func writeInitialReplicaState(
 	gcThreshold hlc.Timestamp,
 	txnSpanGCThreshold hlc.Timestamp,
 ) (enginepb.MVCCStats, error) {
-	rsl := stateloader.Make(st, desc.RangeID)
+	rsl := Make(st, desc.RangeID)
 
 	var s storagebase.ReplicaState
 	s.TruncatedState = &roachpb.RaftTruncatedState{
@@ -104,11 +106,11 @@ func writeInitialReplicaState(
 	return newMS, nil
 }
 
-// writeInitialState calls writeInitialReplicaState followed by
-// synthesizeRaftState. It is typically called during bootstrap. The supplied
+// WriteInitialState calls WriteInitialReplicaState followed by
+// SynthesizeRaftState. It is typically called during bootstrap. The supplied
 // MVCCStats are used for the Stats field after adjusting for persisting the
 // state itself, and the updated stats are returned.
-func writeInitialState(
+func WriteInitialState(
 	ctx context.Context,
 	st *cluster.Settings,
 	eng engine.ReadWriter,
@@ -118,95 +120,12 @@ func writeInitialState(
 	gcThreshold hlc.Timestamp,
 	txnSpanGCThreshold hlc.Timestamp,
 ) (enginepb.MVCCStats, error) {
-	newMS, err := writeInitialReplicaState(ctx, st, eng, ms, desc, lease, gcThreshold, txnSpanGCThreshold)
+	newMS, err := WriteInitialReplicaState(ctx, st, eng, ms, desc, lease, gcThreshold, txnSpanGCThreshold)
 	if err != nil {
 		return enginepb.MVCCStats{}, err
 	}
-	if err := stateloader.Make(st, desc.RangeID).SynthesizeRaftState(ctx, eng); err != nil {
+	if err := Make(st, desc.RangeID).SynthesizeRaftState(ctx, eng); err != nil {
 		return enginepb.MVCCStats{}, err
 	}
 	return newMS, nil
-}
-
-// ClusterSettings returns the node's ClusterSettings.
-func (r *Replica) ClusterSettings() *cluster.Settings {
-	return r.store.cfg.Settings
-}
-
-// In-memory state, immutable fields, and debugging methods are accessed directly.
-
-// StoreID returns the Replica's StoreID.
-func (r *Replica) StoreID() roachpb.StoreID {
-	return r.store.StoreID()
-}
-
-// EvalKnobs returns the EvalContext's Knobs.
-func (r *Replica) EvalKnobs() batcheval.TestingKnobs {
-	return r.store.cfg.TestingKnobs.EvalKnobs
-}
-
-// Tracer returns the Replica's Tracer.
-func (r *Replica) Tracer() opentracing.Tracer {
-	return r.store.Tracer()
-}
-
-// Clock returns the hlc clock shared by this replica.
-func (r *Replica) Clock() *hlc.Clock {
-	return r.store.Clock()
-}
-
-// DB returns the Replica's client DB.
-func (r *Replica) DB() *client.DB {
-	return r.store.DB()
-}
-
-// Engine returns the Replica's underlying Engine. In most cases the
-// evaluation Batch should be used instead.
-func (r *Replica) Engine() engine.Engine {
-	return r.store.Engine()
-}
-
-// AbortSpan returns the Replica's AbortSpan.
-func (r *Replica) AbortSpan() *abortspan.AbortSpan {
-	// Despite its name, the AbortSpan doesn't hold on-disk data in
-	// memory. It just provides methods that take a Batch, so SpanSet
-	// declarations are enforced there.
-	return r.abortSpan
-}
-
-// GetLimiters returns the Replica's limiters.
-func (r *Replica) GetLimiters() *batcheval.Limiters {
-	return &r.store.limiters
-}
-
-// GetTxnWaitQueue returns the Replica's txnwait.Queue.
-func (r *Replica) GetTxnWaitQueue() *txnwait.Queue {
-	return r.txnWaitQueue
-}
-
-// GetTerm returns the term of the given index in the raft log.
-func (r *Replica) GetTerm(i uint64) (uint64, error) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return r.raftTermRLocked(i)
-}
-
-// GetRangeID returns the Range ID.
-func (r *Replica) GetRangeID() roachpb.RangeID {
-	return r.RangeID
-}
-
-// GetGCThreshold returns the GC threshold.
-func (r *Replica) GetGCThreshold() hlc.Timestamp {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return *r.mu.state.GCThreshold
-}
-
-// GetTxnSpanGCThreshold returns the time of the replica's last transaction span
-// GC.
-func (r *Replica) GetTxnSpanGCThreshold() hlc.Timestamp {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return *r.mu.state.TxnSpanGCThreshold
 }
